requirements.txt
python-dotenv
numpy
tqdm
faiss-cpu
requests
